+++ src/main/java/mediathek/gui/bookmark/BookmarkDialog.java
@@ -28,7 +28,6 @@ import ca.odell.glazedlists.swing.GlazedListsSwing;
 import ca.odell.glazedlists.swing.TableComparatorChooser;
 import mediathek.config.Daten;
 import mediathek.controller.history.SeenHistoryController;
-import mediathek.gui.bookmark.dialog.BookmarkEditNoteDialog;
 import mediathek.gui.bookmark.renderer.*;
 import mediathek.gui.tabs.tab_film.FilmDescriptionPanel;
 import mediathek.mainwindow.MediathekGui;